+++ sem/plotting/tapas_sem_predicted_delta_plot.m
@@ -1,0 +1,15 @@
+function tapas_sem_predicted_delta_plot(time, congruent, incongruent, ...
+    quantiles)
+%% Plot the predicted delta plots 
+%
+% Input
+%
+% Output
+%
+
+% aponteeduardo@gmail.com
+% copyright (C) 2019
+%
+
+end
+